Move nested frames examples to sink
